+++ pkg/resumable/state.go
@@ -1,44 +1,37 @@
 package resumable
 
 import (
+	"encoding/binary"
 	"encoding/json"
 	"fmt"
-	"os"
-	"path"
-	"strconv"
-	"strings"
-	"sync"
-
+	"github.com/Altinity/clickhouse-backup/v2/pkg/common"
 	"github.com/rs/zerolog/log"
+	bolt "go.etcd.io/bbolt"
+	"math/big"
+	"path"
 )
 
+var bucketName = []byte("clickhouse-backup")
+
 type State struct {
-	stateFile    string
-	currentState string
-	params       map[string]interface{}
-	fp           *os.File
-	mx           *sync.RWMutex
+	stateFile string
+	db        *bolt.DB
+	params    map[string]interface{}
 }
 
 func NewState(defaultDiskPath, backupName, command string, params map[string]interface{}) *State {
 	s := State{
-		stateFile:    path.Join(defaultDiskPath, "backup", backupName, fmt.Sprintf("%s.state", command)),
-		currentState: "",
-		mx:           &sync.RWMutex{},
+		stateFile: path.Join(defaultDiskPath, "backup", backupName, fmt.Sprintf("%s.state2", command)),
 	}
-	fp, err := os.OpenFile(s.stateFile, os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0644)
-	if err != nil && !os.IsNotExist(err) {
+	db, err := bolt.Open(s.stateFile, 0600, nil)
+	if err != nil {
 		log.Warn().Msgf("can't open %s error: %v", s.stateFile, err)
+		return nil
 	}
-	s.fp = fp
+	s.db = db
 	s.LoadState()
 	s.LoadParams()
-	if len(s.params) == 0 && params != nil {
-		s.params = params
-		if paramsBytes, err := json.Marshal(s.params); err == nil {
-			s.AppendToState(string(paramsBytes), 0)
-		}
-	}
+	s.CleanupStateIfParamsChange(params)
 	return &s
 }
 
@@ -46,74 +39,147 @@ func (s *State) GetParams() map[string]interface{} {
 	return s.params
 }
 
+func (s *State) getBucket(tx *bolt.Tx) *bolt.Bucket {
+	bucket := tx.Bucket(bucketName)
+	if bucket == nil {
+		log.Fatal().Msgf("can't open bucket %s in %s", bucketName, s.stateFile)
+	}
+	return bucket
+}
+
 func (s *State) LoadParams() {
-	lines := strings.SplitN(s.currentState, "\n", 2)
-	if len(lines) == 0 || !strings.HasPrefix(lines[0], "{") {
+	if s.db == nil {
 		return
 	}
-	//size 0 during write
-	lines[0] = strings.TrimSuffix(lines[0], ":0")
-	if err := json.Unmarshal([]byte(lines[0]), &s.params); err != nil {
-		log.Error().Msgf("can't parse state file line 0 as []interface{}: %s", lines[0])
+	if err := s.db.View(func(tx *bolt.Tx) error {
+		bucket := s.getBucket(tx)
+		params := bucket.Get([]byte("params"))
+		if params != nil {
+			s.params = make(map[string]interface{})
+			return json.Unmarshal(params, &s.params)
+		}
+		return nil
+	}); err != nil {
+		log.Warn().Msgf("can't load params from %s, error: %v", s.stateFile, err)
 	}
 }
 
 func (s *State) LoadState() {
-	s.mx.Lock()
-	state, err := os.ReadFile(s.stateFile)
-	if err == nil {
-		s.currentState = string(state)
-	} else {
-		s.currentState = ""
-		if !os.IsNotExist(err) {
-			log.Warn().Msgf("can't read %s error: %v", s.stateFile, err)
-		} else {
-			log.Warn().Msgf("%s empty, will continue from scratch error: %v", s.stateFile, err)
+	if s.db == nil {
+		return
+	}
+	err := s.db.Update(func(tx *bolt.Tx) error {
+		var err error
+		bucket := tx.Bucket(bucketName)
+		if bucket == nil {
+			bucket, err = tx.CreateBucket(bucketName)
+			if err != nil {
+				return fmt.Errorf("can't create bucket: %s", err)
+			}
+		}
+		return nil
+	})
+	if err != nil {
+		log.Warn().Msgf("LoadState error: %v", err)
+	}
+}
+
+func (s *State) CleanupStateIfParamsChange(params map[string]interface{}) {
+	needCleanup := false
+	if s.params != nil && params == nil {
+		needCleanup = true
+	}
+	if s.params != nil && params != nil && !common.CompareMaps(s.params, params) {
+		log.Debug().Msgf("SUKA s.params != nil %v && params != nil %v", s.params != nil, params != nil)
+		needCleanup = true
+	}
+
+	if needCleanup {
+		log.Info().Msgf("parameters changed old=%#v new=%#v, %s cleanup begin", s.params, params, s.stateFile)
+		err := s.db.Batch(func(tx *bolt.Tx) error {
+			b := s.getBucket(tx)
+			c := b.Cursor()
+			for k, _ := c.First(); k != nil; k, _ = c.Next() {
+				if err := b.Delete(k); err != nil {
+					return err
+				}
+			}
+			return nil
+		})
+		if err != nil {
+			log.Warn().Msgf("can't cleanupBucket %s in %s", bucketName, s.stateFile)
 		}
 	}
-	s.mx.Unlock()
+	_ = s.db.Batch(func(tx *bolt.Tx) error {
+		b := s.getBucket(tx)
+		s.saveParams(b, params)
+		return nil
+	})
+
+}
+
+func (s *State) saveParams(b *bolt.Bucket, params map[string]interface{}) {
+	if params != nil {
+		s.params = params
+	}
+	if s.params == nil {
+		return
+	}
+	paramsBytes, err := json.Marshal(s.params)
+	if err != nil {
+		log.Warn().Msgf("can't json.Marshal(s.params=%#v): %v", s.params, err)
+		return
+	}
+	if err = b.Put([]byte("params"), paramsBytes); err != nil {
+		log.Warn().Msgf("can't bolt.Put(s.params): %v", err)
+	}
 }
 
 func (s *State) AppendToState(path string, size int64) {
-	path = fmt.Sprintf("%s:%d", path, size)
-	s.mx.Lock()
-	if s.fp != nil {
-		_, err := s.fp.WriteString(path + "\n")
-		if err != nil {
-			log.Warn().Msgf("can't write %s error: %v", s.stateFile, err)
-		}
-		err = s.fp.Sync()
-		if err != nil {
-			log.Warn().Msgf("can't sync %s error: %v", s.stateFile, err)
-		}
+	if s.db == nil {
+		return
 	}
-	s.currentState += path + "\n"
-	s.mx.Unlock()
+	err := s.db.Update(func(tx *bolt.Tx) error {
+		b := s.getBucket(tx)
+		buf := make([]byte, binary.MaxVarintLen64)
+		n := binary.PutVarint(buf, size)
+		return b.Put([]byte(path), buf[:n])
+	})
+	if err != nil {
+		log.Fatal().Msgf("can't write key %s to %s error: %v", path, s.stateFile, err)
+	}
 }
 
 func (s *State) IsAlreadyProcessedBool(path string) bool {
 	isProcesses, _ := s.IsAlreadyProcessed(path)
 	return isProcesses
 }
+
 func (s *State) IsAlreadyProcessed(path string) (bool, int64) {
-	var size int64
-	var err error
-	s.mx.RLock()
-	res := strings.Index(s.currentState, path+":")
-	if res >= 0 {
-		// s.logger is non thread-safe https://github.com/rs/zerolog/issues/242
-		log.Info().Msgf("%s already processed", path)
-		sSize := s.currentState[res : res+strings.Index(s.currentState[res:], "\n")]
-		sSize = sSize[strings.Index(sSize, ":")+1:]
-		size, err = strconv.ParseInt(sSize, 10, 64)
-		if err != nil {
-			log.Warn().Msgf("invalid size %s in upload state: %v", sSize, err)
-		}
+	if s.db == nil {
+		return false, 0
 	}
-	s.mx.RUnlock()
-	return res >= 0, size
+	size := int64(0)
+	found := false
+	err := s.db.View(func(tx *bolt.Tx) error {
+		b := s.getBucket(tx)
+		buf := b.Get([]byte(path))
+		if buf != nil {
+			found = true
+			size = new(big.Int).SetBytes(buf).Int64()
+			log.Info().Msgf("%s already processed", path)
+		}
+		return nil
+	})
+	if err != nil {
+		log.Fatal().Msgf("can't read key %s to %s error: %v", path, s.stateFile, err)
+		return false, 0
+	}
+	return found, size
 }
 
 func (s *State) Close() {
-	_ = s.fp.Close()
+	if err := s.db.Close(); err != nil {
+		log.Warn().Err(err).Msgf("can't close %s", s.stateFile)
+	}
 }